backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
pypdf==3.17.1
openai==1.54.0
httpx==0.27.0
python-dotenv==1.0.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.3
faiss-cpu==1.7.4
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.24.3